refactor: extract useReminderEngine hook
Moves activeReminders state, firedRemindersRef, and lastReminderDateRef
out of App.jsx into a dedicated useReminderEngine hook. Also moves the
reminder scheduling useEffect, auto-dismiss useEffect, native Android
alarm-sync useEffect, and the snoozeReminder / dismissReminder /
dismissAllReminders functions. Pure helpers (getTaskCategory,
getReminderPoints, timeToMinutes, minutesToTime) become module-level
functions inside the hook file.
